Refactor metadata handler structure
