--- internal/handlers/metadata.go
@@ -1,914 +1,0 @@
-package handlers
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"reflect"
-	"sort"
-	"strings"
-	"sync"
-
-	"github.com/nlstn/go-odata/internal/metadata"
-	"github.com/nlstn/go-odata/internal/response"
-)
-
-// MetadataHandler handles metadata document requests
-type MetadataHandler struct {
-	entities map[string]*metadata.EntityMetadata
-	// Cached metadata documents
-	cachedXML  string
-	cachedJSON []byte
-	onceXML    sync.Once
-	onceJSON   sync.Once
-	namespace  string
-}
-
-const defaultNamespace = "ODataService"
-
-// NewMetadataHandler creates a new metadata handler
-func NewMetadataHandler(entities map[string]*metadata.EntityMetadata) *MetadataHandler {
-	return &MetadataHandler{
-		entities:  entities,
-		namespace: defaultNamespace,
-	}
-}
-
-// SetNamespace updates the namespace used for metadata generation and clears cached documents.
-func (h *MetadataHandler) SetNamespace(namespace string) {
-	trimmed := strings.TrimSpace(namespace)
-	if trimmed == "" {
-		trimmed = defaultNamespace
-	}
-	if trimmed == h.namespace {
-		return
-	}
-	h.namespace = trimmed
-	h.cachedXML = ""
-	h.cachedJSON = nil
-	h.onceXML = sync.Once{}
-	h.onceJSON = sync.Once{}
-}
-
-func (h *MetadataHandler) namespaceOrDefault() string {
-	if strings.TrimSpace(h.namespace) == "" {
-		return defaultNamespace
-	}
-	return h.namespace
-}
-
-func (h *MetadataHandler) qualifiedTypeName(typeName string) string {
-	return fmt.Sprintf("%s.%s", h.namespaceOrDefault(), typeName)
-}
-
-type enumTypeInfo struct {
-	Members        []metadata.EnumMember
-	IsFlags        bool
-	UnderlyingType string
-}
-
-// HandleMetadata handles the metadata document endpoint
-func (h *MetadataHandler) HandleMetadata(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case http.MethodGet, http.MethodHead:
-		h.handleGetMetadata(w, r)
-	case http.MethodOptions:
-		h.handleOptionsMetadata(w)
-	default:
-		if err := response.WriteError(w, http.StatusMethodNotAllowed, "Method not allowed",
-			fmt.Sprintf("Method %s is not supported for metadata document", r.Method)); err != nil {
-			fmt.Printf("Error writing error response: %v\n", err)
-		}
-	}
-}
-
-// handleGetMetadata handles GET requests for metadata document
-func (h *MetadataHandler) handleGetMetadata(w http.ResponseWriter, r *http.Request) {
-	// Check if JSON format is requested via $format query parameter or Accept header
-	useJSON := shouldReturnJSON(r)
-
-	if useJSON {
-		h.handleMetadataJSON(w, r)
-	} else {
-		h.handleMetadataXML(w, r)
-	}
-}
-
-// handleOptionsMetadata handles OPTIONS requests for metadata document
-func (h *MetadataHandler) handleOptionsMetadata(w http.ResponseWriter) {
-	w.Header().Set("Allow", "GET, HEAD, OPTIONS")
-	w.WriteHeader(http.StatusOK)
-}
-
-// shouldReturnJSON determines if JSON format should be returned based on request
-func shouldReturnJSON(r *http.Request) bool {
-	// Check $format query parameter first (highest priority)
-	format := r.URL.Query().Get("$format")
-	if format == "json" || format == "application/json" {
-		return true
-	}
-	if format == "xml" || format == "application/xml" {
-		return false
-	}
-
-	// Check Accept header with proper content negotiation
-	accept := r.Header.Get("Accept")
-	if accept == "" {
-		// No Accept header - default to XML for metadata per OData v4 spec
-		return false
-	}
-
-	// Parse Accept header and find the best match
-	// Handle cases like:
-	// - "application/json"
-	// - "application/xml"
-	// - "application/json;q=0.9, application/xml;q=0.8"
-	// - "*/*"
-	// - "text/html, application/xml;q=0.9, */*;q=0.8"
-
-	type mediaType struct {
-		mimeType string
-		quality  float64
-	}
-
-	parts := strings.Split(accept, ",")
-	mediaTypes := make([]mediaType, 0, len(parts))
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-		if part == "" {
-			continue
-		}
-
-		// Split by semicolon to separate media type from parameters
-		subparts := strings.Split(part, ";")
-		mimeType := strings.TrimSpace(subparts[0])
-		quality := 1.0 // Default quality
-
-		// Parse quality value if present
-		for _, param := range subparts[1:] {
-			param = strings.TrimSpace(param)
-			if strings.HasPrefix(param, "q=") {
-				if q, err := parseQuality(param[2:]); err == nil {
-					quality = q
-				}
-			}
-		}
-
-		mediaTypes = append(mediaTypes, mediaType{mimeType: mimeType, quality: quality})
-	}
-
-	// Find the best matching media type
-	var bestJSON, bestXML, bestWildcard float64
-	for _, mt := range mediaTypes {
-		switch mt.mimeType {
-		case "application/json":
-			if mt.quality > bestJSON {
-				bestJSON = mt.quality
-			}
-		case "application/xml", "text/xml":
-			if mt.quality > bestXML {
-				bestXML = mt.quality
-			}
-		case "*/*", "application/*":
-			if mt.quality > bestWildcard {
-				bestWildcard = mt.quality
-			}
-		}
-	}
-
-	// If both JSON and XML are explicitly specified, choose the one with higher quality
-	// If qualities are equal, prefer JSON (tie-breaking rule)
-	if bestJSON > 0 && bestXML > 0 {
-		return bestJSON >= bestXML
-	}
-
-	// If only JSON is specified (not via wildcard), return JSON
-	if bestJSON > 0 {
-		return true
-	}
-
-	// If only XML is specified, return XML
-	if bestXML > 0 {
-		return false
-	}
-
-	// If only wildcard is specified, default to XML for metadata
-	if bestWildcard > 0 {
-		return false
-	}
-
-	// Default to XML for metadata per OData v4 spec
-	return false
-}
-
-// parseQuality parses a quality value from Accept header
-func parseQuality(s string) (float64, error) {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return 1.0, nil
-	}
-	// Simple parsing - just handle common cases
-	switch s {
-	case "1", "1.0", "1.00", "1.000":
-		return 1.0, nil
-	case "0.9":
-		return 0.9, nil
-	case "0.8":
-		return 0.8, nil
-	case "0.7":
-		return 0.7, nil
-	case "0.6":
-		return 0.6, nil
-	case "0.5":
-		return 0.5, nil
-	case "0":
-		return 0.0, nil
-	default:
-		// Try to parse as float
-		var q float64
-		_, err := fmt.Sscanf(s, "%f", &q)
-		if err != nil {
-			return 1.0, err
-		}
-		// Quality must be between 0 and 1
-		if q < 0 {
-			q = 0
-		}
-		if q > 1 {
-			q = 1
-		}
-		return q, nil
-	}
-}
-
-// handleMetadataXML handles XML metadata format (existing implementation)
-func (h *MetadataHandler) handleMetadataXML(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/xml")
-	w.WriteHeader(http.StatusOK)
-
-	// For HEAD requests, don't write the body
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	// Use sync.Once to ensure metadata is built only once
-	h.onceXML.Do(func() {
-		h.cachedXML = h.buildMetadataDocument()
-	})
-
-	if _, err := w.Write([]byte(h.cachedXML)); err != nil {
-		fmt.Printf("Error writing metadata response: %v\n", err)
-	}
-}
-
-// buildMetadataDocument builds the complete metadata XML document
-func (h *MetadataHandler) buildMetadataDocument() string {
-	namespace := h.namespaceOrDefault()
-	metadata := fmt.Sprintf(`<?xml version="1.0" encoding="UTF-8"?>
-<edmx:Edmx xmlns:edmx="http://docs.oasis-open.org/odata/ns/edmx" Version="%s">
-  <edmx:DataServices>
-    <Schema xmlns="http://docs.oasis-open.org/odata/ns/edm" Namespace="%s">
-`, response.ODataVersionValue, namespace)
-
-	// Add enum types
-	metadata += h.buildEnumTypes()
-
-	// Add entity types
-	metadata += h.buildEntityTypes()
-
-	// Add entity container with entity sets
-	metadata += h.buildEntityContainer()
-
-	metadata += `    </Schema>
-  </edmx:DataServices>
-</edmx:Edmx>`
-
-	return metadata
-}
-
-// buildEnumTypes builds the enum type definitions
-func (h *MetadataHandler) buildEnumTypes() string {
-	enumDefinitions := make(map[string]*enumTypeInfo)
-
-	for _, entityMeta := range h.entities {
-		for _, prop := range entityMeta.Properties {
-			if !prop.IsEnum || prop.EnumTypeName == "" {
-				continue
-			}
-
-			info, exists := enumDefinitions[prop.EnumTypeName]
-			if !exists {
-				info = &enumTypeInfo{}
-				enumDefinitions[prop.EnumTypeName] = info
-			}
-
-			if len(info.Members) == 0 && len(prop.EnumMembers) > 0 {
-				info.Members = append([]metadata.EnumMember(nil), prop.EnumMembers...)
-			}
-			if info.UnderlyingType == "" && prop.EnumUnderlyingType != "" {
-				info.UnderlyingType = prop.EnumUnderlyingType
-			}
-			if prop.IsFlags {
-				info.IsFlags = true
-			}
-		}
-	}
-
-	if len(enumDefinitions) == 0 {
-		return ""
-	}
-
-	enumNames := make([]string, 0, len(enumDefinitions))
-	for name := range enumDefinitions {
-		enumNames = append(enumNames, name)
-	}
-	sort.Strings(enumNames)
-
-	var builder strings.Builder
-	for _, name := range enumNames {
-		info := enumDefinitions[name]
-		if info == nil || len(info.Members) == 0 {
-			continue
-		}
-		builder.WriteString(h.buildEnumType(name, info))
-	}
-
-	return builder.String()
-}
-
-// buildEnumType builds a single enum type definition
-func (h *MetadataHandler) buildEnumType(enumTypeName string, info *enumTypeInfo) string {
-	if info == nil {
-		return ""
-	}
-
-	flagsAttr := ""
-	if info.IsFlags {
-		flagsAttr = ` IsFlags="true"`
-	}
-
-	underlyingType := "Edm.Int32"
-	if info.UnderlyingType != "" {
-		underlyingType = info.UnderlyingType
-	}
-
-	var builder strings.Builder
-	builder.WriteString(fmt.Sprintf(`      <EnumType Name="%s" UnderlyingType="%s"%s>`, enumTypeName, underlyingType, flagsAttr))
-	builder.WriteString("\n")
-
-	for _, member := range info.Members {
-		builder.WriteString(fmt.Sprintf(`        <Member Name="%s" Value="%d" />`, member.Name, member.Value))
-		builder.WriteString("\n")
-	}
-
-	builder.WriteString("      </EnumType>\n")
-	return builder.String()
-}
-
-// buildEntityTypes builds the entity type definitions
-func (h *MetadataHandler) buildEntityTypes() string {
-	result := ""
-	for _, entityMeta := range h.entities {
-		result += h.buildEntityType(entityMeta)
-	}
-	return result
-}
-
-// buildEntityType builds a single entity type definition
-func (h *MetadataHandler) buildEntityType(entityMeta *metadata.EntityMetadata) string {
-	// Add HasStream attribute for media entities
-	hasStreamAttr := ""
-	if entityMeta.HasStream {
-		hasStreamAttr = ` HasStream="true"`
-	}
-
-	result := fmt.Sprintf(`      <EntityType Name="%s"%s>
-        <Key>
-`, entityMeta.EntityName, hasStreamAttr)
-
-	// Add all key properties (supports composite keys)
-	for _, keyProp := range entityMeta.KeyProperties {
-		result += fmt.Sprintf(`          <PropertyRef Name="%s" />
-`, keyProp.JsonName)
-	}
-
-	result += `        </Key>
-`
-
-	// Add regular properties
-	result += h.buildRegularProperties(entityMeta)
-
-	// Add navigation properties
-	result += h.buildNavigationProperties(entityMeta)
-
-	result += `      </EntityType>
-`
-	return result
-}
-
-// buildRegularProperties builds the regular (non-navigation) properties
-func (h *MetadataHandler) buildRegularProperties(entityMeta *metadata.EntityMetadata) string {
-	result := ""
-	for _, prop := range entityMeta.Properties {
-		if prop.IsNavigationProp {
-			continue // Handle navigation properties separately
-		}
-
-		// Skip stream content fields as they are accessed through the stream property
-		if prop.IsStream || strings.HasSuffix(prop.FieldName, "ContentType") || strings.HasSuffix(prop.FieldName, "Content") {
-			// Check if this field is part of a stream property
-			isStreamField := false
-			for _, streamProp := range entityMeta.StreamProperties {
-				if prop.FieldName == streamProp.StreamContentTypeField || prop.FieldName == streamProp.StreamContentField {
-					isStreamField = true
-					break
-				}
-			}
-			if isStreamField {
-				continue
-			}
-		}
-
-		// Determine the EDM type
-		var edmType string
-		if prop.IsEnum && prop.EnumTypeName != "" {
-			edmType = h.qualifiedTypeName(prop.EnumTypeName)
-		} else {
-			edmType = getEdmType(prop.Type)
-		}
-
-		// Determine nullable
-		nullable := "false"
-		if prop.Nullable != nil {
-			if *prop.Nullable {
-				nullable = "true"
-			} else {
-				nullable = "false"
-			}
-		} else if !prop.IsRequired && !prop.IsKey {
-			nullable = "true"
-		}
-
-		// Build property attributes
-		attrs := fmt.Sprintf(`Name="%s" Type="%s" Nullable="%s"`, prop.JsonName, edmType, nullable)
-
-		// Add facets
-		if prop.MaxLength > 0 {
-			attrs += fmt.Sprintf(` MaxLength="%d"`, prop.MaxLength)
-		}
-		if prop.Precision > 0 {
-			attrs += fmt.Sprintf(` Precision="%d"`, prop.Precision)
-		}
-		if prop.Scale > 0 {
-			attrs += fmt.Sprintf(` Scale="%d"`, prop.Scale)
-		}
-		if prop.DefaultValue != "" {
-			attrs += fmt.Sprintf(` DefaultValue="%s"`, prop.DefaultValue)
-		}
-
-		result += fmt.Sprintf(`        <Property %s />
-`, attrs)
-	}
-
-	// Add stream properties
-	for _, streamProp := range entityMeta.StreamProperties {
-		result += fmt.Sprintf(`        <Property Name="%s" Type="Edm.Stream" />
-`, streamProp.Name)
-	}
-
-	return result
-}
-
-// buildNavigationProperties builds the navigation properties
-func (h *MetadataHandler) buildNavigationProperties(entityMeta *metadata.EntityMetadata) string {
-	result := ""
-	for _, prop := range entityMeta.Properties {
-		if !prop.IsNavigationProp {
-			continue
-		}
-
-		typeName := h.qualifiedTypeName(prop.NavigationTarget)
-		if prop.NavigationIsArray {
-			typeName = fmt.Sprintf("Collection(%s)", typeName)
-		}
-
-		// Check if we have referential constraints
-		if len(prop.ReferentialConstraints) > 0 {
-			result += fmt.Sprintf(`        <NavigationProperty Name="%s" Type="%s">
-`, prop.JsonName, typeName)
-			result += `          <ReferentialConstraint>
-`
-			for dependent, principal := range prop.ReferentialConstraints {
-				result += fmt.Sprintf(`            <Property Name="%s" ReferencedProperty="%s" />
-`, dependent, principal)
-			}
-			result += `          </ReferentialConstraint>
-        </NavigationProperty>
-`
-		} else {
-			result += fmt.Sprintf(`        <NavigationProperty Name="%s" Type="%s" />
-`, prop.JsonName, typeName)
-		}
-	}
-	return result
-}
-
-// buildEntityContainer builds the entity container with entity sets and singletons
-func (h *MetadataHandler) buildEntityContainer() string {
-	result := `      <EntityContainer Name="Container">
-`
-	// Build entity sets and singletons
-	for entitySetName, entityMeta := range h.entities {
-		if entityMeta.IsSingleton {
-			// Output singleton definition
-			result += fmt.Sprintf(`        <Singleton Name="%s" Type="%s">
-`, entityMeta.SingletonName, h.qualifiedTypeName(entityMeta.EntityName))
-
-			// Add navigation property bindings for singleton
-			for _, prop := range entityMeta.Properties {
-				if prop.IsNavigationProp {
-					targetEntitySet := pluralize(prop.NavigationTarget)
-					result += fmt.Sprintf(`          <NavigationPropertyBinding Path="%s" Target="%s" />
-`, prop.JsonName, targetEntitySet)
-				}
-			}
-
-			result += `        </Singleton>
-`
-		} else {
-			// Output entity set definition
-			result += fmt.Sprintf(`        <EntitySet Name="%s" EntityType="%s">
-`, entitySetName, h.qualifiedTypeName(entityMeta.EntityName))
-
-			// Add navigation property bindings
-			for _, prop := range entityMeta.Properties {
-				if prop.IsNavigationProp {
-					targetEntitySet := pluralize(prop.NavigationTarget)
-					result += fmt.Sprintf(`          <NavigationPropertyBinding Path="%s" Target="%s" />
-`, prop.JsonName, targetEntitySet)
-				}
-			}
-
-			result += `        </EntitySet>
-`
-		}
-	}
-
-	result += `      </EntityContainer>
-`
-	return result
-}
-
-// handleMetadataJSON handles JSON metadata format (CSDL JSON)
-func (h *MetadataHandler) handleMetadataJSON(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-
-	// For HEAD requests, don't write the body
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	// Use sync.Once to ensure metadata is built only once
-	h.onceJSON.Do(func() {
-		h.cachedJSON = h.buildMetadataJSON()
-	})
-
-	if _, err := w.Write(h.cachedJSON); err != nil {
-		fmt.Printf("Error writing JSON metadata response: %v\n", err)
-	}
-}
-
-// buildMetadataJSON builds the JSON metadata document and returns the raw bytes
-func (h *MetadataHandler) buildMetadataJSON() []byte {
-	// Build CSDL JSON structure
-	namespace := h.namespaceOrDefault()
-	odataService := make(map[string]interface{})
-	csdl := map[string]interface{}{
-		"$Version":         response.ODataVersionValue,
-		"$EntityContainer": fmt.Sprintf("%s.Container", namespace),
-	}
-	csdl[namespace] = odataService
-
-	// Add enum types
-	h.addJSONEnumTypes(odataService)
-
-	// Add entity types
-	for _, entityMeta := range h.entities {
-		entityType := h.buildJSONEntityType(entityMeta)
-		odataService[entityMeta.EntityName] = entityType
-	}
-
-	// Add entity container
-	container := h.buildJSONEntityContainer()
-	odataService["Container"] = container
-
-	// Encode to JSON bytes
-	jsonBytes, err := json.MarshalIndent(csdl, "", "  ")
-	if err != nil {
-		fmt.Printf("Error marshaling JSON metadata: %v\n", err)
-		return []byte("{}")
-	}
-
-	return jsonBytes
-}
-
-// addJSONEnumTypes adds enum type definitions to the JSON service
-func (h *MetadataHandler) addJSONEnumTypes(odataService map[string]interface{}) {
-	enumDefinitions := make(map[string]*enumTypeInfo)
-
-	for _, entityMeta := range h.entities {
-		for _, prop := range entityMeta.Properties {
-			if !prop.IsEnum || prop.EnumTypeName == "" {
-				continue
-			}
-
-			info, exists := enumDefinitions[prop.EnumTypeName]
-			if !exists {
-				info = &enumTypeInfo{}
-				enumDefinitions[prop.EnumTypeName] = info
-			}
-
-			if len(info.Members) == 0 && len(prop.EnumMembers) > 0 {
-				info.Members = append([]metadata.EnumMember(nil), prop.EnumMembers...)
-			}
-			if info.UnderlyingType == "" && prop.EnumUnderlyingType != "" {
-				info.UnderlyingType = prop.EnumUnderlyingType
-			}
-			if prop.IsFlags {
-				info.IsFlags = true
-			}
-		}
-	}
-
-	if len(enumDefinitions) == 0 {
-		return
-	}
-
-	enumNames := make([]string, 0, len(enumDefinitions))
-	for name := range enumDefinitions {
-		enumNames = append(enumNames, name)
-	}
-	sort.Strings(enumNames)
-
-	for _, name := range enumNames {
-		info := enumDefinitions[name]
-		if info == nil || len(info.Members) == 0 {
-			continue
-		}
-		enumType := h.buildJSONEnumType(info)
-		odataService[name] = enumType
-	}
-}
-
-// buildJSONEnumType builds a JSON enum type definition
-func (h *MetadataHandler) buildJSONEnumType(info *enumTypeInfo) map[string]interface{} {
-	if info == nil {
-		return nil
-	}
-
-	enumType := map[string]interface{}{
-		"$Kind": "EnumType",
-	}
-
-	underlyingType := "Edm.Int32"
-	if info.UnderlyingType != "" {
-		underlyingType = info.UnderlyingType
-	}
-	enumType["$UnderlyingType"] = underlyingType
-
-	if info.IsFlags {
-		enumType["$IsFlags"] = true
-	}
-
-	for _, member := range info.Members {
-		enumType[member.Name] = member.Value
-	}
-
-	return enumType
-}
-
-// buildJSONEntityType builds a JSON entity type definition
-func (h *MetadataHandler) buildJSONEntityType(entityMeta *metadata.EntityMetadata) map[string]interface{} {
-	entityType := make(map[string]interface{})
-	entityType["$Kind"] = "EntityType"
-
-	// Add key(s) - supports both single and composite keys
-	keyNames := make([]string, 0, len(entityMeta.KeyProperties))
-	for _, keyProp := range entityMeta.KeyProperties {
-		keyNames = append(keyNames, keyProp.JsonName)
-	}
-	entityType["$Key"] = keyNames
-
-	// Add regular properties
-	h.addJSONRegularProperties(entityType, entityMeta)
-
-	// Add navigation properties
-	h.addJSONNavigationProperties(entityType, entityMeta)
-
-	return entityType
-}
-
-// addJSONRegularProperties adds regular properties to a JSON entity type
-func (h *MetadataHandler) addJSONRegularProperties(entityType map[string]interface{}, entityMeta *metadata.EntityMetadata) {
-	for _, prop := range entityMeta.Properties {
-		if prop.IsNavigationProp {
-			continue // Handle navigation properties separately
-		}
-
-		propDef := h.buildJSONPropertyDefinition(&prop)
-		entityType[prop.JsonName] = propDef
-	}
-}
-
-// buildJSONPropertyDefinition builds a JSON property definition
-func (h *MetadataHandler) buildJSONPropertyDefinition(prop *metadata.PropertyMetadata) map[string]interface{} {
-	propDef := make(map[string]interface{})
-
-	// Determine the type
-	if prop.IsEnum && prop.EnumTypeName != "" {
-		propDef["$Type"] = h.qualifiedTypeName(prop.EnumTypeName)
-	} else {
-		propDef["$Type"] = getEdmType(prop.Type)
-	}
-
-	// Set nullable
-	if prop.Nullable != nil {
-		propDef["$Nullable"] = *prop.Nullable
-	} else if !prop.IsRequired && !prop.IsKey {
-		propDef["$Nullable"] = true
-	}
-
-	// Add facets
-	h.addJSONPropertyFacets(propDef, prop)
-
-	return propDef
-}
-
-// addJSONPropertyFacets adds facets to a JSON property definition
-func (h *MetadataHandler) addJSONPropertyFacets(propDef map[string]interface{}, prop *metadata.PropertyMetadata) {
-	if prop.MaxLength > 0 {
-		propDef["$MaxLength"] = prop.MaxLength
-	}
-	if prop.Precision > 0 {
-		propDef["$Precision"] = prop.Precision
-	}
-	if prop.Scale > 0 {
-		propDef["$Scale"] = prop.Scale
-	}
-	if prop.DefaultValue != "" {
-		propDef["$DefaultValue"] = prop.DefaultValue
-	}
-}
-
-// addJSONNavigationProperties adds navigation properties to a JSON entity type
-func (h *MetadataHandler) addJSONNavigationProperties(entityType map[string]interface{}, entityMeta *metadata.EntityMetadata) {
-	for _, prop := range entityMeta.Properties {
-		if !prop.IsNavigationProp {
-			continue
-		}
-
-		navProp := h.buildJSONNavigationProperty(&prop)
-		entityType[prop.JsonName] = navProp
-	}
-}
-
-// buildJSONNavigationProperty builds a JSON navigation property definition
-func (h *MetadataHandler) buildJSONNavigationProperty(prop *metadata.PropertyMetadata) map[string]interface{} {
-	navProp := make(map[string]interface{})
-	navProp["$Kind"] = "NavigationProperty"
-
-	if prop.NavigationIsArray {
-		navProp["$Collection"] = true
-		navProp["$Type"] = h.qualifiedTypeName(prop.NavigationTarget)
-	} else {
-		navProp["$Type"] = h.qualifiedTypeName(prop.NavigationTarget)
-	}
-
-	// Add referential constraints if present
-	if len(prop.ReferentialConstraints) > 0 {
-		constraints := make([]map[string]string, 0, len(prop.ReferentialConstraints))
-		for dependent, principal := range prop.ReferentialConstraints {
-			constraints = append(constraints, map[string]string{
-				"Property":           dependent,
-				"ReferencedProperty": principal,
-			})
-		}
-		navProp["$ReferentialConstraint"] = constraints
-	}
-
-	return navProp
-}
-
-// buildJSONEntityContainer builds the JSON entity container with entity sets and singletons
-func (h *MetadataHandler) buildJSONEntityContainer() map[string]interface{} {
-	container := map[string]interface{}{
-		"$Kind": "EntityContainer",
-	}
-
-	for entitySetName, entityMeta := range h.entities {
-		if entityMeta.IsSingleton {
-			// Build singleton definition
-			singleton := map[string]interface{}{
-				"$Type": h.qualifiedTypeName(entityMeta.EntityName),
-			}
-
-			// Add navigation property bindings
-			navigationBindings := h.buildNavigationBindings(entityMeta)
-			if len(navigationBindings) > 0 {
-				singleton["$NavigationPropertyBinding"] = navigationBindings
-			}
-
-			container[entityMeta.SingletonName] = singleton
-		} else {
-			// Build entity set definition
-			entitySet := map[string]interface{}{
-				"$Collection": true,
-				"$Type":       h.qualifiedTypeName(entityMeta.EntityName),
-			}
-
-			// Add navigation property bindings
-			navigationBindings := h.buildNavigationBindings(entityMeta)
-			if len(navigationBindings) > 0 {
-				entitySet["$NavigationPropertyBinding"] = navigationBindings
-			}
-
-			container[entitySetName] = entitySet
-		}
-	}
-
-	return container
-}
-
-// buildNavigationBindings builds navigation property bindings for an entity
-func (h *MetadataHandler) buildNavigationBindings(entityMeta *metadata.EntityMetadata) map[string]string {
-	navigationBindings := make(map[string]string)
-	for _, prop := range entityMeta.Properties {
-		if prop.IsNavigationProp {
-			targetEntitySet := pluralize(prop.NavigationTarget)
-			navigationBindings[prop.JsonName] = targetEntitySet
-		}
-	}
-	return navigationBindings
-}
-
-// getEdmType converts a Go type to an EDM (Entity Data Model) type
-func getEdmType(goType reflect.Type) string {
-	// Handle pointer types
-	if goType.Kind() == reflect.Ptr {
-		goType = goType.Elem()
-	}
-
-	// Check for specific types by name
-	typeName := goType.String()
-	switch typeName {
-	case "time.Time":
-		return "Edm.DateTimeOffset"
-	case "uuid.UUID", "github.com/google/uuid.UUID":
-		return "Edm.Guid"
-	}
-
-	switch goType.Kind() {
-	case reflect.String:
-		return "Edm.String"
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32:
-		return "Edm.Int32"
-	case reflect.Int64:
-		return "Edm.Int64"
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32:
-		return "Edm.Int32"
-	case reflect.Uint64:
-		return "Edm.Int64"
-	case reflect.Float32:
-		return "Edm.Single"
-	case reflect.Float64:
-		return "Edm.Double"
-	case reflect.Bool:
-		return "Edm.Boolean"
-	default:
-		// Check for arrays of bytes (often used for binary data)
-		if goType.Kind() == reflect.Slice && goType.Elem().Kind() == reflect.Uint8 {
-			return "Edm.Binary"
-		}
-		// For complex types, return as string
-		return "Edm.String"
-	}
-}
-
-// pluralize creates a simple pluralized form of the entity name
-func pluralize(word string) string {
-	if word == "" {
-		return word
-	}
-
-	// Simple pluralization rules
-	switch {
-	case strings.HasSuffix(word, "y"):
-		return word[:len(word)-1] + "ies"
-	case strings.HasSuffix(word, "s") || strings.HasSuffix(word, "x") || strings.HasSuffix(word, "z") ||
-		strings.HasSuffix(word, "ch") || strings.HasSuffix(word, "sh"):
-		return word + "es"
-	default:
-		return word + "s"
-	}
-}
